# swagger.json
{
    "openapi": "3.0.0",
    "servers": [
      {
        "url": "http://localhost:5000/api/v1/",
        "description": "Devlopment server (uses test data)",
        "variables": {}
      },
      {
        "url": "http://15.185.219.194/api/v1/",
        "description": "Production server (uses live data)",
        "variables": {},
        "basePath": ""
      }
    ],
    "info": {
      "version": "v1",
      "title": "BOSTA ASSESSMENT API",
      "description": "This api was developed  by eng : Ahmed Ali Ahmed Abd Al Mowla ✌️✌️ ,for BOST company, @copyright BOSTA 2021",
      "termsOfService": "",
      "contact": {},
      "license": { "name": "" }
    },
...
    "paths": {
...
      "/user/resend/confirmation/email": {
        "post": {
          "summary": "Resend confirmation email.",
          "description": "you can use this end point to resend confirmation email.",
          "operationId": "Resend confirmation email",
          "responses": { "200": { "description": "", "headers": {} } },
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "email": "ahmedshekh@student.aast.edu"
                  }
                },
                "example": "{\r\n    \"email\":\"ahmedshekh@student.aast.edu\"\r\n}"
              }
            }
          },
          "tags": ["User"]
        }
        
      },
      "/check": {
        "get": {
          "summary": "Get all cheks",
          "operationId": "you can use this end point to get all checks.",
          "parameters": [
          ],
          "responses": { "200": { "description": "", "headers": {} } },
          "tags": ["Check"]
        },
        "post": {
          "summary": "Create new check",
          "description": "you can use this end point to create new check. optional [ port, webhook, interval, timeout, threshold, authUserName, authPassword, assertStatusCode ]",
          "operationId": "Create new check",
          "responses": { "200": { "description": "", "headers": {} } },
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "name": "npm2",
                    "url": "google",
                    "protocal": "HTTPS",
                    "path": "/",
                    "authUserName": "ahmed",
                    "authPassword": "ahmed123",
                    "webhook": "https://slack.com/545a5d4a5",
                    "interval": "1",
                    "timeout": "4000",
                    "threshold": "3",
                    "assertStatusCode": "200"
                  }
                },
                "example": "{\r\n    \"name\":\"google\",\r\n    \"url\":\"google.com\",\r\n    \"protocal\":\"HTTPS\",\r\n    \"path\":\"/\",\r\n  \"port\":\"\",\r\n  \"authUserName\":\"ahmed\",\r\n  \"authPassword\":\"ahmed123\",\r\n    \"webhook\":\"https://slack.com/545a5d4a5\",\r\n    \"interval\":\"1\",\r\n    \"timeout\":\"4000\",\r\n    \"threshold\":\"3\",\r\n    \"assertStatusCode\":\"200\"\r\n}"
              }
            }
          },
          "tags": ["Check"]
        }
        
      },
...
    },
    "components": {
...
        "securitySchemes": {
          "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-auth-token",
            "scheme": "bearer",
            "description": "Enter JWT token",
            "in": "header"
          }
        }
      },
      "security": [{ "ApiKeyAuth": [] }],
...
}